fix: harden security validation and deduplicate logic
