+++ .changeset/harden-security-validation.md
@@ -1,0 +1,5 @@
+---
+"@googleworkspace/cli": patch
+---
+
+Security: Harden validate_resource_name and fix Gmail watch path traversal
