+++ src/helpers/mod.rs
@@ -63,156 +63,3 @@ pub fn get_helper(service: &str) -> Option<Box<dyn Helper>> {
         _ => None,
     }
 }
-
-// ---------------------------------------------------------------------------
-// URL safety helpers
-// ---------------------------------------------------------------------------
-
-/// Percent-encode a value for use as a single URL path segment (e.g., file ID,
-/// calendar ID, message ID). All non-alphanumeric characters are encoded.
-pub(crate) fn encode_path_segment(s: &str) -> String {
-    use percent_encoding::{utf8_percent_encode, NON_ALPHANUMERIC};
-    utf8_percent_encode(s, NON_ALPHANUMERIC).to_string()
-}
-
-/// Validate a multi-segment resource name (e.g., `spaces/ABC`, `subscriptions/123`).
-/// Rejects path traversal and control characters while preserving the intentional
-/// `/`-delimited structure. Returns the validated name or an error with a clear
-/// message suitable for LLM callers.
-pub(crate) fn validate_resource_name(s: &str) -> Result<&str, GwsError> {
-    if s.is_empty() {
-        return Err(GwsError::Validation(
-            "Resource name must not be empty".to_string(),
-        ));
-    }
-    if s.split('/').any(|seg| seg == "..") {
-        return Err(GwsError::Validation(format!(
-            "Resource name must not contain path traversal ('..') segments: {s}"
-        )));
-    }
-    if s.contains('\0') || s.chars().any(|c| c.is_control()) {
-        return Err(GwsError::Validation(format!(
-            "Resource name contains invalid characters: {s}"
-        )));
-    }
-    // Reject URL-special characters that could inject query params or fragments
-    if s.contains('?') || s.contains('#') {
-        return Err(GwsError::Validation(format!(
-            "Resource name must not contain '?' or '#': {s}"
-        )));
-    }
-    Ok(s)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // -- encode_path_segment --------------------------------------------------
-
-    #[test]
-    fn test_encode_path_segment_plain_id() {
-        assert_eq!(encode_path_segment("abc123"), "abc123");
-    }
-
-    #[test]
-    fn test_encode_path_segment_email() {
-        // Calendar IDs are often email addresses
-        let encoded = encode_path_segment("user@gmail.com");
-        assert!(!encoded.contains('@'));
-        assert!(!encoded.contains('.'));
-    }
-
-    #[test]
-    fn test_encode_path_segment_query_injection() {
-        // LLM might include query params in an ID by mistake
-        let encoded = encode_path_segment("fileid?fields=name");
-        assert!(!encoded.contains('?'));
-        assert!(!encoded.contains('='));
-    }
-
-    #[test]
-    fn test_encode_path_segment_fragment_injection() {
-        let encoded = encode_path_segment("fileid#section");
-        assert!(!encoded.contains('#'));
-    }
-
-    #[test]
-    fn test_encode_path_segment_path_traversal() {
-        // Encoding makes traversal segments harmless
-        let encoded = encode_path_segment("../../etc/passwd");
-        assert!(!encoded.contains('/'));
-        assert!(!encoded.contains(".."));
-    }
-
-    #[test]
-    fn test_encode_path_segment_unicode() {
-        // LLM might pass unicode characters
-        let encoded = encode_path_segment("日本語ID");
-        assert!(!encoded.contains('日'));
-    }
-
-    #[test]
-    fn test_encode_path_segment_spaces() {
-        let encoded = encode_path_segment("my file id");
-        assert!(!encoded.contains(' '));
-    }
-
-    #[test]
-    fn test_encode_path_segment_already_encoded() {
-        // LLM might double-encode by passing pre-encoded values
-        let encoded = encode_path_segment("user%40gmail.com");
-        // The % itself gets encoded to %25, so %40 becomes %2540
-        // This prevents double-encoding issues at the HTTP layer
-        assert!(encoded.contains("%2540"));
-    }
-
-    // -- validate_resource_name -----------------------------------------------
-
-    #[test]
-    fn test_validate_resource_name_valid() {
-        assert!(validate_resource_name("spaces/ABC123").is_ok());
-        assert!(validate_resource_name("subscriptions/my-sub").is_ok());
-        assert!(validate_resource_name("@default").is_ok());
-        assert!(validate_resource_name("projects/p1/topics/t1").is_ok());
-    }
-
-    #[test]
-    fn test_validate_resource_name_traversal() {
-        assert!(validate_resource_name("../../etc/passwd").is_err());
-        assert!(validate_resource_name("spaces/../other").is_err());
-        assert!(validate_resource_name("..").is_err());
-    }
-
-    #[test]
-    fn test_validate_resource_name_control_chars() {
-        assert!(validate_resource_name("spaces/\0bad").is_err());
-        assert!(validate_resource_name("spaces/\nbad").is_err());
-        assert!(validate_resource_name("spaces/\rbad").is_err());
-        assert!(validate_resource_name("spaces/\tbad").is_err());
-    }
-
-    #[test]
-    fn test_validate_resource_name_empty() {
-        assert!(validate_resource_name("").is_err());
-    }
-
-    #[test]
-    fn test_validate_resource_name_query_injection() {
-        // LLMs might append query strings or fragments to resource names
-        assert!(validate_resource_name("spaces/ABC?key=val").is_err());
-        assert!(validate_resource_name("spaces/ABC#fragment").is_err());
-    }
-
-    #[test]
-    fn test_validate_resource_name_error_messages_are_clear() {
-        let err = validate_resource_name("").unwrap_err();
-        assert!(err.to_string().contains("must not be empty"));
-
-        let err = validate_resource_name("../bad").unwrap_err();
-        assert!(err.to_string().contains("path traversal"));
-
-        let err = validate_resource_name("bad\0id").unwrap_err();
-        assert!(err.to_string().contains("invalid characters"));
-    }
-}
