+++ src/helpers/workflows.rs
@@ -383,7 +383,7 @@ async fn handle_meeting_prep(matches: &ArgMatches) -> Result<(), GwsError> {
 
     let events_url = format!(
         "https://www.googleapis.com/calendar/v3/calendars/{}/events",
-        crate::helpers::encode_path_segment(calendar_id),
+        crate::validate::encode_path_segment(calendar_id),
     );
     let events_json = get_json(
         &client,
@@ -459,7 +459,7 @@ async fn handle_email_to_task(matches: &ArgMatches) -> Result<(), GwsError> {
     // 1. Fetch the email
     let msg_url = format!(
         "https://gmail.googleapis.com/gmail/v1/users/me/messages/{}",
-        crate::helpers::encode_path_segment(message_id),
+        crate::validate::encode_path_segment(message_id),
     );
     let msg_json = get_json(
         &client,
@@ -495,7 +495,7 @@ async fn handle_email_to_task(matches: &ArgMatches) -> Result<(), GwsError> {
         "notes": format!("From email: {}\n\n{}", message_id, snippet),
     });
 
-    let tasklist = crate::helpers::validate_resource_name(tasklist)?;
+    let tasklist = crate::validate::validate_resource_name(tasklist)?;
     let task_url = format!(
         "https://tasks.googleapis.com/tasks/v1/lists/{}/tasks",
         tasklist,
@@ -628,7 +628,7 @@ async fn handle_file_announce(matches: &ArgMatches) -> Result<(), GwsError> {
     // 1. Fetch file metadata from Drive
     let file_url = format!(
         "https://www.googleapis.com/drive/v3/files/{}",
-        crate::helpers::encode_path_segment(file_id),
+        crate::validate::encode_path_segment(file_id),
     );
     let file_json = get_json(
         &client,
@@ -653,7 +653,7 @@ async fn handle_file_announce(matches: &ArgMatches) -> Result<(), GwsError> {
         .unwrap_or_else(|| format!("📎 {file_name}\n{file_link}"));
 
     let chat_body = json!({ "text": msg_text });
-    let space = crate::helpers::validate_resource_name(space)?;
+    let space = crate::validate::validate_resource_name(space)?;
     let chat_url = format!("https://chat.googleapis.com/v1/{}/messages", space);
 
     let chat_resp = client
